+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased - 2018-07-04
+
+### Added
+- Implemented VocaLink V5.10 specification.
+
 ## [v1.7.0] - 2018-07-04
 
 ### Added